requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
autograd==1.5
backcall==0.2.0
biopython==1.81
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.6
click==8.1.3
cloudpickle==2.2.1
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
diffrax==0.2.2
distrax==0.1.3
dm-haiku==0.0.9
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.10
einops==0.6.0
entrypoints==0.4
equinox==0.9.2
etils==1.0.0
executing==0.8.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.5.3
geomstats==2.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
hydra-colorlog==1.1.0
hydra-core==1.1.1
hydra-joblib-launcher==1.1.5
hydra-submitit-launcher==1.2.0.dev1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.19.2
ipython==8.10.0
jax==0.3.15
jaxlib==0.3.15+cuda11.cudnn805
jaxtyping==0.2.13
jedi==0.18.1
jmp==0.0.4
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.2.0
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.1.2
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
score-sde==1.0.0
seaborn==0.12.2
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
submitit==1.4.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-probability==0.19.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.1
torch-tb-profiler==0.4.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.38.4
yaml-include==1.0.0
yarl==1.8.2
zipp==3.15.0
